Move the test to the common trait
